requirements.txt
python-telegram-bot==21.10
anthropic==0.43.0
supabase==2.11.0
apscheduler==3.10.4
python-dotenv==1.0.1
